Revert "Revert "Refactor Delete Command Class.""
This reverts commit 54ab55acc36221690619340697c3a98fcd79318d.

	renamed:    src/main/java/seedu/address/logic/commands/DeleteCommand.java -> src/main/java/seedu/address/logic/commands/client/DeleteClientCommand.java
	renamed:    src/main/java/seedu/address/logic/parser/DeleteCommandParser.java -> src/main/java/seedu/address/logic/parser/DeleteClientCommandParser.java
	renamed:    src/test/java/seedu/address/logic/parser/DeleteCommandParserTest.java -> src/test/java/seedu/address/logic/parser/DeleteClientCommandParserTest.java
