+++ src/main/java/seedu/address/logic/commands/client/DeleteClientCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.client;
 
 import static java.util.Objects.requireNonNull;
 
@@ -6,6 +6,8 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.client.Client;
@@ -13,9 +15,9 @@ import seedu.address.model.client.Client;
 /**
  * Deletes a Client identified using it's displayed index from the address book.
  */
-public class DeleteCommand extends Command {
+public class DeleteClientCommand extends Command {
 
-    public static final String COMMAND_WORD = "delete";
+    public static final String COMMAND_WORD = "cdel";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD
             + ": Deletes the Client identified by the index number used in the displayed Client list.\n"
